Pass the secret name, rather than sharing a const
To be able to refer to a Secret in a Stack, my initial instinct was to
use the same const as the secret's name and as the reference. However:
Pulumi will usually treat the supplied name as a prefix and generate a
name for resources, so to avoid different invocations colliding, so
this won't work.

Instead, within the same module, I can create the secret then use its
(lifted) name property as the reference. This is good enough for now;
for sharing it across modules, or across programs, I think it'll need
to be referenced via an output.

Signed-off-by: Michael Bridgen <mbridgen@pulumi.com>