Pass the bootstrapped names via config
This makes a few tweaks to pass info from the bootstrap program to the
(very similar) self-syncing program. In particular, the secret name
and repo URL are put in the stack's config, where they can be fished
out and used for the Kubernetes object, _and_ put in the stack's
config recursively.

Signed-off-by: Michael Bridgen <mbridgen@pulumi.com>