Do CRD install and stack naming just once
I've failed to build a model of how CRDs work with Pulumi, so I'm giving
up and fudging it by making them part of the bootstrap program rather
than part of the operator component. For some reason that I would love
to understand, this gets past the problem of it always doing the wrong
thing with CRDs.

In a similar vein, I've not figured out how to do a "handover" from one
stack to another, which was my intention with the operator deployment
and so on. Possibly I just have the wrong idea about how this should
work. In any case, this commit makes it so that the stack name used for
the bootstrap stage becomes the name of the automated stack, as defined
in the Stack custom resource.

Signed-off-by: Michael Bridgen <mbridgen@pulumi.com>